+++ app/src/main/java/ngo/teog/swift/gui/ImageActivity.java
@@ -24,6 +24,7 @@ import java.security.NoSuchAlgorithmException;
 import ngo.teog.swift.R;
 import ngo.teog.swift.communication.RequestFactory;
 import ngo.teog.swift.communication.VolleyManager;
+import ngo.teog.swift.gui.imageCapture.ImageCaptureActivity;
 import ngo.teog.swift.helpers.Defaults;
 import ngo.teog.swift.helpers.ResourceKeys;
